backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
alpaca-py==0.25.2
python-dotenv==1.0.0
cryptography==42.0.0
requests==2.31.0
websocket-client==1.7.0
pandas==2.1.4
numpy==1.26.3
schedule==1.2.0
pytz==2024.1